+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # domino x.x.x (not yet released)
+* Bug fix for `Element#closest` when selector doesn't match (#154)
 
 # domino 2.1.3 (6 Mar 2019)
 * Bug fix for CSS `$=` selector and for matches on root `<html>` element.